homework-4: correcting hw4

+++ src/com/cursor/hw4/task2/abstracts/Product.java
@@ -4,7 +4,8 @@ import com.cursor.hw4.task2.interfaces.Buyable;
 
 public abstract class Product implements Buyable {
     String name;
-    int price;
-    int quantity;
-    int ageRestriction;
+    double price;
+    double quantity;
+    double ageRestriction;
+
 }
